Start writing Umbrel app parser
The goal is to have an exports.sh parser that translates it to things available on Citadel as efficiently as possible

+++ src/composegenerator/umbrel/types.rs
@@ -2,12 +2,12 @@
 use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Eq)]
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
 #[cfg_attr(feature = "schema", derive(JsonSchema))]
 #[serde(rename_all = "camelCase")]
 pub struct Metadata {
     /// The version of the metadata file
-    pub manifest_version: i8,
+    pub manifest_version: f64,
     /// The app id
     pub id: String,
     /// The name of the app
